requirements.txt
# Core Infrastructure
livekit==0.12.0
livekit-agents==0.8.0
asyncio-mqtt==0.16.1
websockets==12.0

# Deep Learning & Computer Vision
torch>=2.6.0
torchvision>=0.21.0
opencv-python>=4.9.0
mediapipe>=0.10.9
hsemotion>=0.2.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
silero-vad==4.0.0
nemo-toolkit[asr]==1.22.0  # For SenseVoice
transformers==4.37.2  # For Wav2Vec2

# LLM & NLP
llama-cpp-python==0.2.55  # For Llama 3.1 with CUDA support
sentence-transformers==2.3.1
tiktoken==0.6.0

# TTS
TTS==0.22.0  # Coqui TTS base
requests==2.31.0

# Data & Utilities
numpy>=1.26.0
pandas>=2.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Database (for memory)
redis==5.0.1
motor==3.3.2  # Async MongoDB
pymongo==4.6.1

# Security & Privacy
cryptography==42.0.2
python-jose[cryptography]==3.3.0

# API & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
locust==2.20.0  # For load testing

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0
